# 02-development/homework/balancio/openapi.yaml
openapi: 3.0.3
info:
  title: Balancio API
  description: >
    Shared expense splitter for friends, households, couples, trips, and
    other small groups. See _docs/specs.md for the full product spec.
  version: "0.1.0"
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{user_id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      summary: Get a user by id
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [users]
      summary: Update a user's display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/reset-password:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [users]
      summary: Reset a user's password to a new temporary one (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/force-password-change:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [users]
      summary: Force a user to change their password on next login (admin only)
      responses:
        "204":
          description: Flag set
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/deactivate:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [users]
      summary: Deactivate a user (admin only, requires zero global balance)
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: User has a non-zero global balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/activate:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [users]
      summary: Reactivate a deactivated user (admin only)
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /settlements:
    get:
      tags: [settlements]
      summary: List settlements
      responses:
        "200":
          description: List of settlements, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      tags: [settlements]
      summary: Create a settlement (pending, awaiting recipient confirmation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSettlementRequest"
      responses:
        "201":
          description: Settlement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          $ref: "#/components/responses/BadRequest"
  /settlements/{settlement_id}:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    patch:
      tags: [settlements]
      summary: Edit a pending settlement (payer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettlementRequest"
      responses:
        "200":
          description: Updated settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          description: Settlement is not pending, or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /settlements/{settlement_id}/confirm:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    post:
      tags: [settlements]
      summary: Confirm a pending settlement (recipient only)
      responses:
        "200":
          description: Confirmed settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /settlements/{settlement_id}/reject:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    post:
      tags: [settlements]
      summary: Reject a pending settlement (recipient only)
      responses:
        "200":
          description: Rejected settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /settlements/{settlement_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    post:
      tags: [settlements]
      summary: Cancel a pending settlement (payer only)
      responses:
        "200":
          description: Cancelled settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /settlements/{settlement_id}/request-reversal:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    post:
      tags: [settlements]
      summary: Request reversal of a confirmed settlement (either participant)
      responses:
        "200":
          description: Settlement now reversal_pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          description: Settlement is not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /settlements/{settlement_id}/confirm-reversal:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    post:
      tags: [settlements]
      summary: Confirm a pending reversal (the other participant)
      responses:
        "200":
          description: Settlement now reversed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          description: No pending reversal, or same user who requested it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /settlements/{settlement_id}/reject-reversal:
    parameters:
      - $ref: "#/components/parameters/SettlementId"
    post:
      tags: [settlements]
      summary: Reject a pending reversal (settlement returns to confirmed)
      responses:
        "200":
          description: Settlement back to confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          description: No pending reversal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
...
    SettlementId:
      name: settlement_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation error or invalid financial state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Friendly, domain-specific message. Never a stack trace.
    Role:
      type: string
      enum: [admin, user]
    Theme:
      type: string
      enum: [light, dark]
    Money:
      type: string
      description: Decimal amount with exactly two decimal places.
      pattern: '^\d+\.\d{2}$'
      example: "123.45"
    User:
      type: object
      required:
        - id
        - username
        - display_name
        - role
        - is_active
        - must_change_password
        - theme
        - created_at
        - updated_at
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        is_active:
          type: boolean
        must_change_password:
          type: boolean
        theme:
          $ref: "#/components/schemas/Theme"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateUserRequest:
      type: object
      required: [username, display_name, temporary_password]
      properties:
        username:
          type: string
        display_name:
          type: string
        temporary_password:
          type: string
          minLength: 8
    UpdateUserRequest:
      type: object
      properties:
        display_name:
          type: string
    ResetPasswordRequest:
      type: object
      required: [new_temporary_password]
      properties:
        new_temporary_password:
          type: string
          minLength: 8
...
    SettlementStatus:
      type: string
      enum:
        - pending
        - confirmed
        - rejected
        - cancelled
        - reversal_pending
        - reversed
    Settlement:
      type: object
      required:
        - id
        - payer_id
        - recipient_id
        - amount
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        payer_id:
          type: string
        recipient_id:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        note:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/SettlementStatus"
        reversal_requested_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateSettlementRequest:
      type: object
      required: [payer_id, recipient_id, amount]
      properties:
        payer_id:
          type: string
        recipient_id:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        note:
          type: string
    UpdateSettlementRequest:
      type: object
      properties:
        amount:
          $ref: "#/components/schemas/Money"
        note:
          type: string
...